requirements.txt
alembic==1.8.0
async-timeout==4.0.2
cachelib==0.7.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-Caching==1.11.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
redis==4.3.3
requests==2.27.1
SQLAlchemy==1.4.37
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
